+++ src/main.rs
@@ -62,7 +62,7 @@ fn main() {
         }
     }
     history::init(&mut rl);
-    rl.set_completer(Rc::new(completers::CicadaCompleter));
+    rl.set_completer(Rc::new(completers::CicadaCompleter{sh: Rc::new(sh.clone())}));
 
     let mut status = 0;
     loop {